requirements.txt
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
shap==0.44.0
matplotlib==3.8.2
seaborn==0.13.1
jupyter==1.0.0
notebook==7.0.6
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
psutil==5.9.8
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==6.1.0
bcrypt==4.0.1
python-multipart==0.0.9
paho-mqtt==1.6.1